Fixed a bug with data filtering
Data filtering was applied when plotting a part of the genome. However, the additional parameters (such as cex, pch, col, labels...) were not filtered accordingly, matching in some situactions data points with the wrong parameters. I've added a filterParams function to utils.R and updated all plotting functions to make use of it to correctly filter the parameters. A side effect is that plotting functions are now call with do.call. Plotting functions producing a single large object (kpLines, kpRibbon, etc...) did not need an update since assigning a different lwd or cex to different data.points does not make sense in this context.